+++ src/Http/Controllers/LicenseController.php
@@ -68,8 +68,8 @@ class LicenseController extends Controller
             }
         }
 
-        $this->installer->completeStep(4);
-        $this->installer->setCurrentStep(5);
+        $this->installer->completeStep(5);
+        $this->installer->setCurrentStep(6);
 
         return redirect()->route('installer.admin');
     }